graspit + mrroc++ project
git-svn-id: https://segomo.elka.pw.edu.pl/svn/mrrocpp/base/trunk@4103 4c39dba3-d726-0410-8099-c98d187771e6

+++ src/ecp_mp/transmitter/transmitter.h
@@ -67,7 +67,7 @@ typedef struct _FROM_VA
 			char status;
 		} draughts_ai;
 		struct {
-			float grasp_joint[7];
+			double grasp_joint[7];
 		} graspit;
 		playerc_joystick_t player_joystick;
 		playerc_position_t player_position;
